Prevent the program from hanging due to rounding
When the number of rounds was not a multiple of the number of threads, or when the number of rounds was less than the percentage allocated to each thread and etc, the progress bar would never reach 100 due to rounding. As a result the program would wait endlessly and never print a result. To fix this, explicitly set the progress bar to 100 when we are done - this is a much simpler solution than using FP math to accurately handle all cases and the progress bar is accurate enough anyway.